+++ src/seq/helpers.h
@@ -114,8 +114,8 @@ struct seq_config;
 BARTLIB_API extern int BARTLIB_CALL seq_raga_spokes(const struct seq_config* seq);
 
 // minimum TE and TR calculation at end of prepare
-BARTLIB_API extern long BARTLIB_CALL seq_minimum_tr(const struct seq_config* seq);
-BARTLIB_API extern void BARTLIB_CALL seq_minimum_te(const struct seq_config* seq, long* min_te, long* fil_te);
+BARTLIB_API extern double BARTLIB_CALL seq_minimum_tr(const struct seq_config* seq);
+BARTLIB_API extern void BARTLIB_CALL seq_minimum_te(const struct seq_config* seq, double* min_te, double* fil_te);
 
 
 BARTLIB_API extern long BARTLIB_CALL seq_relevant_readouts_meas_time(const struct seq_config* seq);
